Use new entity naming style
